+++ src/frequenz/sdk/microgrid/_power_distributing/_component_pool_status_tracker.py
@@ -33,6 +33,7 @@ class ComponentPoolStatusTracker:
 
     def __init__(  # pylint: disable=too-many-arguments
         self,
+        *,
         component_ids: abc.Set[int],
         component_status_sender: Sender[ComponentPoolStatus],
         max_data_age: timedelta,