+++ src/frequenz/sdk/timeseries/pv_pool/_pv_pool_reference_store.py
@@ -36,6 +36,7 @@ class PVPoolReferenceStore:
 
     def __init__(  # pylint: disable=too-many-arguments
         self,
+        *,
         channel_registry: ChannelRegistry,
         resampler_subscription_sender: Sender[ComponentMetricRequest],
         status_receiver: Receiver[ComponentPoolStatus],