+++ tests/timeseries/test_periodic_feature_extractor.py
@@ -60,7 +60,9 @@ async def init_feature_extractor_no_data(
     # We only need the moving window to initialize the PeriodicFeatureExtractor class.
     lm_chan = Broadcast[Sample[Quantity]](name="lm_net_power")
     moving_window = MovingWindow(
-        timedelta(seconds=1), lm_chan.new_receiver(), timedelta(seconds=1)
+        size=timedelta(seconds=1),
+        resampled_data_recv=lm_chan.new_receiver(),
+        input_sampling_period=timedelta(seconds=1),
     )
     async with moving_window:
         await lm_chan.new_sender().send(
@@ -126,7 +128,7 @@ async def test_feature_extractor() -> None:  # pylint: disable=too-many-statemen
     async with init_feature_extractor(data, timedelta(seconds=5)) as feature_extractor:
         assert np.allclose(feature_extractor.avg(start, end), [1.5, 1.5])
 
-    async def _test_fun(  # pylint: disable=too-many-arguments
+    async def _test_fun(  # pylint: disable=too-many-arguments,too-many-positional-arguments
         data: list[float],
         period: int,
         start: int,